Demonstrate @LevelBossMike's ember-statecharts-actors demo, but using statechart-component instead.
