--- app/components/providers/posts.hbs
@@ -1,1 +1,0 @@
-{{yield (hash data=this.posts.data fns=this.posts.fns)}}
