--- app/components/providers/posts.js
@@ -1,6 +1,0 @@
-import Component from '@glimmer/component';
-import { inject as service } from '@ember/service';
-
-export default class PostsProvider extends Component {
-  @service posts;
-}
