--- app/routes/index.js
@@ -1,10 +1,0 @@
-import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
-
-export default class IndexRoute extends Route {
-  @service store;
-
-  model() {
-    return this.store.findAll('post', { reload: true });
-  }
-}
