+++ mirage/serializers/application.js
@@ -1,4 +1,3 @@
 import { JSONAPISerializer } from 'ember-cli-mirage';
 
-export default JSONAPISerializer.extend({
-});
+export default JSONAPISerializer.extend({});
