+++ package.json
@@ -39,8 +39,7 @@
     "ember-cli-mirage": "^2.2.0",
     "ember-cli-sri": "^2.1.1",
     "ember-cli-terser": "^4.0.2",
-    "ember-composable-helpers": "^5.0.0",
-    "ember-data": "~4.1.0",
+    "ember-data": "~3.28.0",
     "ember-export-application-global": "^2.0.1",
     "ember-fetch": "^8.1.1",
     "ember-load-initializers": "^2.1.2",
@@ -48,9 +47,9 @@
     "ember-qunit": "^5.1.5",
     "ember-render-helpers": "^0.2.0",
     "ember-resolver": "^8.0.3",
-    "ember-source": "~4.1.0",
-    "ember-statecharts": "^0.14.0-beta.0",
-    "ember-template-lint": "^3.15.0",
+    "ember-source": "~4.2.0-beta.1",
+    "ember-template-imports": "^2.0.1",
+    "ember-template-lint": "^4.0.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-ember": "^10.5.8",
@@ -66,9 +65,20 @@
     "xstate": "^4.27.0"
   },
   "engines": {
-    "node": "12.* || 14.* || >= 16"
+    "node": "14.* || >= 16"
   },
   "ember": {
     "edition": "octane"
+  },
+  "dependencies": {
+    "ember-data-resources": "^2.1.0",
+    "ember-functions-as-helper-polyfill": "^1.0.11",
+    "ember-router-helpers": "^0.4.0",
+    "ember-statechart-component": "^3.2.1"
+  },
+  "volta": {
+    "node": "16.13.1",
+    "yarn": "1.22.17",
+    "npm": "8.3.0"
   }
 }
